FreeCAD: Compiler warning fixes

+++ src/Gui/SoDatumLabel.cpp
@@ -81,7 +81,7 @@ void glDrawLine(const SbVec3f& p1, const SbVec3f& p2){
 void glDrawArc(const SbVec3f& center, float radius, float startAngle=0., float endAngle=2.0*M_PI, int countSegments=0){
     float range = endAngle - startAngle;
     if (range < 0.) {
-        range += 2.0 * M_PI;
+        range += 2.0 * static_cast<float>(M_PI);
     }
 
     if (countSegments == 0){
@@ -549,7 +549,7 @@ private:
         float startangle = atan2f(vc1[1], vc1[0]);
         float endangle = atan2f(vc2[1], vc2[0]);
         if (endangle < startangle) {
-            endangle += 2. * M_PI;
+            endangle += 2. * static_cast<float>(M_PI);
         }
 
         SbVec3f textCenter;
@@ -692,7 +692,7 @@ SbVec3f SoDatumLabel::getLabelTextCenterArcLength(const SbVec3f& ctr, const SbVe
     float endangle = atan2f(vc2[1], vc2[0]);
 
     if (endangle < startangle) {
-        endangle += 2. * M_PI;
+        endangle += 2. * static_cast<float>(M_PI);
     }
 
     // Text location
